Fix type of vat rate var

+++ htdocs/societe/class/societe.class.php
@@ -2482,7 +2482,7 @@ class Societe extends CommonObject
 				$discount->amount_ttc = $discount->multicurrency_amount_ttc = price2num((float) $discount->amount_ht + (float) $discount->amount_tva, 'MT');
 			}
 
-			$discount->tva_tx = price2num($vatrate);
+			$discount->tva_tx = (float) price2num($vatrate);
 			$discount->vat_src_code = $vat_src_code;
 
 			$discount->description = $desc;
